requirements.txt
Django==3.0.7
gunicorn==20.0.4
python-decouple==3.3
pathlib~=1.0.1
python-dotenv~=0.13.0